- rename BackgroundWorkService.kt to WorkService.kt - rename WorkerDatabase.kt to WorkDatabase.kt - rename WorkerDao.kt to WorkDao.kt - rename BootServiceState.kt to WorkServiceState.kt
